Bump Cats to 2.2.0
I ran into a `NoSuchMethodError` in the FS2 runtime regarding `Chain.traverse` into `Seq`, and building this project locally with Cats 2.2.0 seemed to fix the issue. FS2 and Cats Effect are both depending on Cats 2.2.0, while this project depends on 2.1.0 directly. 

I'm not sure if there are other implications of making this change, but figured it would be worth a try.