Switch partitionsMapStream to use per-assignment finisher
The purpose of this change is to allow partition streams returned by
KafkaConsumer.partitionsMapStream to be individually completed, without
having to close the full assignment at once as happened before.

When using the default assignor, there should be no visible change in
behaviour, as partition assignments are revoked in full on a rebalance,
and new assignments sent out to consumers.

When using the CooperativeStickyAssignor, however, partition assignments
may be individually revoked without affecting other assignments. In this
case, partition streams for non-revoked partitions will remain active
until the point they get revoked.

Compared to behaviour observed with the default assignor, when using
partitionsMapStream with CooperativeStickyAssignor the following will be
observed:

- maps in partitionsMapStream will include only entries for newly
  assigned partitions; When partition assignments are revoked, an empty
  map will be emitted.
- partition streams for non-revoked assignments will remain active after
  a new (partial) assignment map is emitted.