Optimise {Committable,}ProducerRecords `apply` methods
Also fixes an unnecessarily tight constraint (`Traverse` can be a
`Foldable`).

**Explanation:**

The `apply` methods on both `CommittableProducerRecords` and
`ProducerRecords` hand-roll an iteration through a `Foldable` in order
to construct an `ArrayBuffer` (initialised using the `size` of the
`Foldable`). This `ArrayBuffer` is then converted to a Scala `Array` and
a `Chunk` instantiated from this.

In the worst-case (I think) this involves 3 iterations through the
`Foldable`:
 - `size`
 - `foldLeft` to construct the `ArrayBuffer`
 - Conversion of `ArrayBuffer` to an `Array`

I think the only step of the current implementation that can be
optimised is the `size` call. For example (I think) `Vector[F].size`
is constant-time (no iteration).

The implementation has been simplified to use `Chunk#iterable` and
`Foldable[F].toIterable` instead. In the best-case this should involve
no iteration as both `toIterable` and `iterable` are optimised (e.g. if
`F` is a `Vector`) . In the worst-case this should involve 1 traversal
(construction of an intermediate `LazyList`).