Change `Blocking` to use `Sync#interruptible`, not `blocking`
This is to avoid being unable to cancel code that is inside blocking
section. At the moment it is possible for user-code to not cancel
immediately, in which case it will block until an internal Kafka
max-block period is reached (by default 1 minute).

Changing to use `interruptible` means the user-code returns immediately
on cancellation, and the underlying operation in Kafka should be
cancelled.

As a concrete example, at the moment producing to a topic which does not
exist blocks until a minute has passed, even if the effect is cancelled.
For example:

```scala
type K
type V

val producer: TransactionalKafkaProducer.WithoutOffsets[IO, Option[K], Option[V]] = ???

producer
  .produceWithoutOffsets(
    ProducerRecords.one(ProducerRecord(topic = "does_not_exist", key = None, value = None))
  )
  .timeout(1.second)
  .unsafeRunSync()
```

Changing to use `interruptible` means it completes after 1 second with
the timeout exception.