Retire FakeFiber
FakeFiber was used by KafkaConsumer to manage:

1. the consumerActor fiber, processing requests (including polls);
2. the pollScheduler fiber, scheduling poll requests (subject to
   backpressure);
3. a fiber combining the above two fibers.

Compared to a regular fiber, FakeFiber offered a method to combine two
fibers by racing them one against the other. The semantics were similar
to the race method for effects, but operating at the fiber level.

In KafkaConsumer, FakeFiber was used with cancellation effects that
returned immediately (i.e., fiber.cancel.start.void). In addition the
fiber outcome was relayed to KafkaConsumer.awaitTermination.

With this change FakeFiber is replaced with an Async[F].race of the
underlying effects. This is managed in the new method
startBackgroundConsumer, which builds upon effects assembled by
runConsumerActor and runPollScheduler. These effects are unwrapped from
the previous fibers.

As before, cancellation of the consumer effect is only waited on in
awaitTermination, where any errors are propagated.

Compared to the original behaviour of FakeFiber.combine, starting with
cats-effect 3.5.0, cancellation of one of the consumer effects will lead
to cancellation of both, as per changes introduced with
https://github.com/typelevel/cats-effect/pull/3453. I'm not sure how
cancelation would come into play here, but the behaviour change looks
appropriate: without one of the racing fibers KafkaConsumer would not be
functional (no polls scheduled, or no requests/polls processed).