Change `Blocking` to use `Sync#interruptible`, not `blocking` (#1126)
This is to avoid being unable to cancel code that is inside blocking
section. At the moment it is possible for user-code to not cancel
immediately, in which case it will block until an internal Kafka
max-block period is reached (by default 1 minute).

Changing to use `interruptible` means the user-code returns immediately
on cancellation, and the underlying operation in Kafka should be
cancelled.